Fix test that is failing due to Dropwizard change
In Dropwizar 2.1.0, the default value of the driverClass field in
DataSourceFactory was changed from an empty string to null.
As a result, our tests broke since they use null-ness to determine
whether the resolution is NONE or PROVIDER_DEFAULT. So, we
need  to change our assertions to expect NONE now, which is
actually more correct, since the driverClass value is not actually
being provided.